Improve FormSubmitFooter
